Fixes for time handling.
git-svn-id: file:///opt/svn/repositories/sonatype.org/nexus/trunk@512 2aa8b3fc-8ebb-4439-a84f-95066eaea8ab

+++ nexus/nexus-app/src/main/java/org/sonatype/nexus/maven/tasks/DefaultSnapshotRemover.java
@@ -209,7 +209,7 @@ public class DefaultSnapshotRemover
 
             if ( days > 0 )
             {
-                this.dateThreshold = System.currentTimeMillis() - ( days * 86400000 );
+                this.dateThreshold = System.currentTimeMillis() - ( (long) days * 86400000L );
             }
             else
             {
