+++ nexus/nexus-proxy/src/main/java/org/sonatype/nexus/proxy/repository/DefaultRepository.java
@@ -379,7 +379,7 @@ public abstract class DefaultRepository
         // else check age
         else
         {
-            return ( ( System.currentTimeMillis() - item.getRemoteChecked() ) > ( maxAge * 60 * 1000 ) );
+            return ( ( System.currentTimeMillis() - item.getRemoteChecked() ) > ( (long) maxAge * 60L * 1000L ) );
         }
     }
